fix(rootfs): refactor database init process
When you complete a recovery of the database for the first time, a new log timeline is started with an ID of 2. When you restore again, the timeline used when the last backup occurred will be replayed. Because of this, if you restored the database and did not perform a backup, all data from that successful recovery will be lost because only WAL logs from the first timeline (the timeline that the database was last backed up) will be restored.

In order to fix this, after completing a database recovery we create a fresh backup in order to establish a new recovery baseline. That way we can now replay from timeline 2.

Other fixes that this refactor includes:

 - database attempting to start halfway through recovery
 - /bin/is_running identifying the database as running during recovery
 - cleaning up database boot script